+++ index.html
@@ -24,7 +24,6 @@
     <!-- Контейнер для изображений от других пользователей -->
     <section class="pictures  container">
       <h2 class="pictures__title  visually-hidden">Фотографии других пользователей</h2>
-
       <!-- Поле для загрузки нового изображения на сайт -->
       <section class="img-upload">
         <div class="img-upload__wrapper">
@@ -231,7 +230,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
   
 </body>
 </html>
