+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "repository": {
-    "type": "git",
+    "type": "module",
     "url": "git@github.com:htmlacademy-univer-javascript-1/2219807-kekstagram-1.git"
   },
   "bugs": {
